refactor: 책 도메인 import 변경 #131
