+++ src/main/java/com/bookripple/api/domain/book/service/BookQueryServiceImpl.java
@@ -1,26 +1,25 @@
 package com.bookripple.api.domain.book.service;
 
-import com.bookripple.api.common.code.BookErrorCode;
-import com.bookripple.api.common.code.MemberErrorCode;
-import com.bookripple.api.common.code.SearchErrorCode;
-import com.bookripple.api.common.error.ApiException;
-import com.bookripple.api.domain.aladin.dto.AladinItemLookUpResDto;
-import com.bookripple.api.domain.aladin.dto.AladinSearchResDto;
-import com.bookripple.api.domain.aladin.service.AladinService;
+import com.bookripple.api.domain.book.code.BookErrorCode;
+import com.bookripple.api.domain.book.code.SearchErrorCode;
 import com.bookripple.api.domain.book.converter.BookConverter;
 import com.bookripple.api.domain.book.dto.BookRes;
 import com.bookripple.api.domain.book.dto.BookSearchRes;
 import com.bookripple.api.domain.book.entity.Book;
 import com.bookripple.api.domain.book.enums.SearchLogType;
 import com.bookripple.api.domain.book.repository.BookRepository;
+import com.bookripple.api.domain.library.enums.LibraryStatus;
+import com.bookripple.api.domain.library.repository.LibraryItemRepository;
+import com.bookripple.api.domain.member.code.MemberErrorCode;
+import com.bookripple.api.domain.member.repository.MemberRepository;
+import com.bookripple.api.domain.reading.repository.ReadingProgressRepository;
+import com.bookripple.api.global.error.ApiException;
+import com.bookripple.api.infrastructure.aladin.dto.AladinItemLookUpResDto;
+import com.bookripple.api.infrastructure.aladin.dto.AladinSearchResDto;
+import com.bookripple.api.infrastructure.aladin.service.AladinService;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Set;
-
-import com.bookripple.api.domain.library.enums.LibraryStatus;
-import com.bookripple.api.domain.library.repository.LibraryItemRepository;
-import com.bookripple.api.domain.member.repository.MemberRepository;
-import com.bookripple.api.domain.reading.repository.ReadingProgressRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -37,11 +36,11 @@ public class BookQueryServiceImpl implements BookQueryService {
   private final ReadingProgressRepository readingProgressRepository;
 
 
-
   //1 알라딘 검색
   @Override
   @Transactional
-  public BookSearchRes searchFromAladin(Long memberId, String keyword, int start, int size, String queryType,
+  public BookSearchRes searchFromAladin(Long memberId, String keyword, int start, int size,
+      String queryType,
       String searchTarget, SearchLogType type) {
     // 최소 검증
     if (keyword == null || keyword.isBlank()) {
@@ -58,24 +57,25 @@ public class BookQueryServiceImpl implements BookQueryService {
     }
 
     AladinSearchResDto resDto =
-            aladinService.search(keyword, start, size, queryType, searchTarget);
+        aladinService.search(keyword, start, size, queryType, searchTarget);
 
     Set<Long> registeredAladinIds = java.util.Collections.emptySet();
 
-    if (memberId != null && resDto != null && resDto.getItem() != null && !resDto.getItem().isEmpty()) {
+    if (memberId != null && resDto != null && resDto.getItem() != null && !resDto.getItem()
+        .isEmpty()) {
       List<Long> aladinIds = resDto.getItem().stream()
-              .map(AladinSearchResDto.Item::getItemId)
-              .filter(id -> id != null)
-              .toList();
+          .map(AladinSearchResDto.Item::getItemId)
+          .filter(id -> id != null)
+          .toList();
 
       List<LibraryStatus> status = List.of(
-              LibraryStatus.COMPLETED,
-              LibraryStatus.READING,
-              LibraryStatus.LIKED
+          LibraryStatus.COMPLETED,
+          LibraryStatus.READING,
+          LibraryStatus.LIKED
       );
 
       registeredAladinIds = new java.util.HashSet<>(
-              libraryItemRepository.findRegisteredAladinBookIds(memberId, status, aladinIds)
+          libraryItemRepository.findRegisteredAladinBookIds(memberId, status, aladinIds)
       );
     }
 
@@ -86,19 +86,20 @@ public class BookQueryServiceImpl implements BookQueryService {
       }
     }
 
-
     return BookConverter.toBookSearchRes(resDto, registeredAladinIds);
   }
 
-    // 추천도서용 알라딘 도서 상세 조회 및 저장
+  // 추천도서용 알라딘 도서 상세 조회 및 저장
   @Override
   @Transactional
   public BookRes getOrCreateByAladinItemId(Long itemId) {
-    if (itemId == null) throw new ApiException(BookErrorCode.NO_BOOK);
+    if (itemId == null) {
+      throw new ApiException(BookErrorCode.NO_BOOK);
+    }
 
     return bookRepository.findByAladinBookId(itemId)
-            .map(BookConverter::toBookRes)
-            .orElseGet(() -> BookConverter.toBookRes(createFromAladinEntity(itemId)));
+        .map(BookConverter::toBookRes)
+        .orElseGet(() -> BookConverter.toBookRes(createFromAladinEntity(itemId)));
 
   }
 
@@ -107,12 +108,16 @@ public class BookQueryServiceImpl implements BookQueryService {
   @Override
   @Transactional
   public BookRes getOrCreateByAladinItemId(Long memberId, Long itemId) {
-    if (itemId == null) throw new ApiException(BookErrorCode.NO_BOOK);
-    if (memberId == null) throw new ApiException(MemberErrorCode.MEMBER_NOT_FOUND);
+    if (itemId == null) {
+      throw new ApiException(BookErrorCode.NO_BOOK);
+    }
+    if (memberId == null) {
+      throw new ApiException(MemberErrorCode.MEMBER_NOT_FOUND);
+    }
 
     // 1) Book 캐시 히트 or 생성
     Book book = bookRepository.findByAladinBookId(itemId)
-            .orElseGet(() -> createFromAladinEntity(itemId));
+        .orElseGet(() -> createFromAladinEntity(itemId));
 
     // 2) 🔥 등록 처리 (항상 실행)
     upsertLibraryItemReading(memberId, book);
@@ -123,31 +128,30 @@ public class BookQueryServiceImpl implements BookQueryService {
   }
 
 
-
   private Book createFromAladinEntity(Long itemId) {
     AladinItemLookUpResDto lookUp = aladinService.lookup(itemId, null);
 
     AladinItemLookUpResDto.Item it =
-            (lookUp == null || lookUp.getItem() == null || lookUp.getItem().isEmpty())
-                    ? null
-                    : lookUp.getItem().get(0);
+        (lookUp == null || lookUp.getItem() == null || lookUp.getItem().isEmpty())
+            ? null
+            : lookUp.getItem().get(0);
 
     if (it == null) {
       throw new ApiException(BookErrorCode.NO_BOOK);
     }
 
     Book book = Book.builder()
-            .aladinBookId(it.getItemId())
-            .title(it.getTitle())
-            .author(it.getAuthor())
-            .publisher(it.getPublisher())
-            .bookCover(it.getCover())
-            .isbn10(it.getIsbn10())
-            .isbn13(it.getIsbn13())
-            .publishedAt(parsePublishedAt(it.getPubDate()))
-            .totalPage(it.getSubInfo() != null ? it.getSubInfo().getItemPage() : null)
-            .story(it.getDescription())
-            .build();
+        .aladinBookId(it.getItemId())
+        .title(it.getTitle())
+        .author(it.getAuthor())
+        .publisher(it.getPublisher())
+        .bookCover(it.getCover())
+        .isbn10(it.getIsbn10())
+        .isbn13(it.getIsbn13())
+        .publishedAt(parsePublishedAt(it.getPubDate()))
+        .totalPage(it.getSubInfo() != null ? it.getSubInfo().getItemPage() : null)
+        .story(it.getDescription())
+        .build();
 
     return bookRepository.save(book);
   }
@@ -173,29 +177,29 @@ public class BookQueryServiceImpl implements BookQueryService {
       itemOpt.get().setStatus(LibraryStatus.READING);
     } else {
       libraryItemRepository.save(
-              com.bookripple.api.domain.library.entity.LibraryItem.builder()
-                      .member(memberRepository.getReferenceById(memberId))
-                      .book(book)
-                      .status(LibraryStatus.READING)
-                      .build()
+          com.bookripple.api.domain.library.entity.LibraryItem.builder()
+              .member(memberRepository.getReferenceById(memberId))
+              .book(book)
+              .status(LibraryStatus.READING)
+              .build()
       );
     }
   }
 
   private void resetReadingProgress(Long memberId, Book book) {
     var progress = readingProgressRepository
-            .findByMemberIdAndBookId(memberId, book.getId());
+        .findByMemberIdAndBookId(memberId, book.getId());
 
     if (progress == null) {
       readingProgressRepository.save(
-              com.bookripple.api.domain.reading.entity.ReadingProgress.builder()
-                      .member(memberRepository.getReferenceById(memberId))
-                      .book(book)
-                      .readingTime(0)
-                      .progress(java.math.BigDecimal.ZERO)
-                      .isLiked(false)
-                      .isCompleted(false)
-                      .build()
+          com.bookripple.api.domain.reading.entity.ReadingProgress.builder()
+              .member(memberRepository.getReferenceById(memberId))
+              .book(book)
+              .readingTime(0)
+              .progress(java.math.BigDecimal.ZERO)
+              .isLiked(false)
+              .isCompleted(false)
+              .build()
       );
     } else {
       progress.setReadingTime(0);
